The raw code for the SVG templates.
If you can't figure out how to open the SVGs as text, here is a plain text file that you can copy. 
All you need to do is:
- Copy the code
- Paste into Notepad
- Replace the hexcodes with the appropriate ones (stripes are in descending order) to color
- Save as .SVG
Now you have an SVG of that flag. You can use https://ezgif.com/svg-to-png to make it into a PNG of any size you wish.