The raw code for the SVG templates.
If you can't figure out how to open the SVGs as text, this is a plain text file that you can copy. 
All you need to do is:
- Copy the code
- Paste into Notepad, or any text editor that allows you to save as a SVG, whether from a menu or from typing .SVG into the filename.
- Replace the hexcodes with the appropriate ones to color; stripes are in descending order
- Save as .SVG / .svg
Now you have an SVG of that flag. You can use https://ezgif.com/svg-to-png to make it into a PNG of any size you wish.
* If you want to make a PNG of a flag on this Github, you can click "Raw" and copy that link, then paste it into that website.